Remove several intermediate copies, just because we can.
.. Also rely on 2018 edition's "? in main" feature to print error
returns.  In particular, this now prints the error (if any) to stderr,
and exits non-zero.

Note, this changes the "public" API of the library to accept a
`String` rather than `&str`, since the code currently requires an
owned copy of the input.  Since this is the public entry point, it
probably deserves a decision around intention/semantics rather than
just what the current implementation does.

Some possible alternatives might be:
- `JWT::new(&str) -> JWT`: Don't keep the full original input after
  `new()` has finished.  Currently the `token` property is private and
  unused.
- `JWT::new(&'a str) -> JWT<'a>`: Keep a borrowed reference to the
  input.